Initial Taxonomy Work for Fingerprinting Devices
hostapd supports device fingerprinting thanks to Denton Gentry. You can read
more on how it works at:
https://media.defcon.org/DEF%20CON%2025/DEF%20CON%2025%20presentations/Denton%20Gentry/DEFCON-25-Denton-Gentry-I-Know-What-You-Are-By-The-Smell-Of-Your-Wifi.pdf

However, it only stored taxonomy info for associated devices. I extended it to
store it for probing devices. Additionally, it will log them to a file along
with SSID, MAC and whether the MAC is random. This will allow devices to be
deanonymised (i.e. different random MACs but similar taxonomy).

This has made the log more dirty with duplicate entries, but that can be fixed
later.