Moved file configs to proper hostapd config options
Previously, we would use environment variables to know what files to
write to. This was a hack. It's now part of the proper hostapd_config.

It's working for log_ssid and mana_outfile. It's also working to log
mschapv2 EAP creds, but needs to be expanded to the other EAPs.

I also finally changed the ennode config option to something that makes
more sense mana_credout.

Two small cleanups too; putting the pre-assoc taxonomy method in the
header and using os_ methods for the ssid filter check.