Extended EAP Logging to MSCHAP, CHAP and PAP
hostapd-mana always logged these types, but the code was messy, and it
turns out the CHAP code was straight wrong. Updated them to use the new
callback method. Also included multiple cracking formats where
appropriate.

Cleaned up the WPE config options.

There looks like a big change in the config code, but I was just
removing an extraneous space that crept in to the start of the line.