Initial WPA/2 Handshake Capture Support
hostapd-mana will now write captured WPA/2 handshakes to a file in the hashcat
hccapx format so they can be cracked.

It currently writes to /tmp/test.hccapx I'll add a configuration for this next.
Handshakes are appended to the file.

This will save needing to run a second device in monitor mode to capture
handshakes in a pcap.