Message when handshake captured & config warning
You'll now see when you captured a handshake.

More importantly, cracking handshakes required an accurate SSID. When in mana
mode we don't know what SSID the station tried to associate to yet. As such,
I've prevented a config with both enable_mana and mana_wpaout set in the
meantime.

Technically, by viewing the context, one could edit the capture file and put in
the right SSID, but that's going to lead to too many github issues. I'd rather
do it right.